KS-3-5 исправлены замечания от 12.06.2025

+++ src/main/kotlin/Lesson_3/lesson_3_task_5.kt
@@ -2,10 +2,10 @@ fun main() {
 
     val moveString = "D2-D4;0"
 
-    val parts = moveString.split("-",";")
-    val from = parts [0]
-    val to = parts [1]
-    val moveNumber = parts[2].toInt()
+    val parts = moveString.split("-", ";")
+    val from = parts[0]
+    val to = parts[1]
+    val moveNumber = parts[2]
 
     println("Откуда: $from")
     println("Куда: $to")
